refactor: change List Component architecture
